// rim-front/src/app/interfaces.ts
export interface InSiteFormJSON{
    id: string;
    fuc: string;
    terminal: string;
    order: string;
    styleButton: string;
    styleBody: string;
    styleBox: string;
    styleBoxText: string;
    buttonValue: string;
    idiomaInsite: string;
    mostrarLogoInsite: string;
    estiloReducidoInsite: string;
    estiloInsite: string;
}

export interface Info {
    Ds_Merchant_MerchantCode: string;
    Ds_Merchant_Terminal: string;
    Ds_Merchant_Order: string;
}